Core Types

##Graph

Missing docstring.

Missing docstring for AbstractCustomNGraph, AbstractCustomNDiGraph. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_graph, get_vtx_ids, get_vtx_map, get_nodes, get_vtx, get_vtx_id, get_node, get_parent, get_edge, replace_edge!, node_val, edge_val, node_id, # may cause an issue edge_source, edge_target, set_vtx_map!, insert_to_vtx_map!, replace_node!, add_node!, make_node, make_edge, add_child!, add_parent!, rem_node!, rem_nodes!. Check Documenter's build log for details.

Tree

GraphUtils.AbstractCustomNTree — Type

abstract type AbstractCustomNTree{N,ID} <: AbstractCustomNDiGraph{N,ID}

Abstract custom graph type with tree edge structure.

Missing docstring.

Missing docstring for AbstractCustomNETree. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractCustomTree. Check Documenter's build log for details.

GraphUtils.AbstractTreeNode — Type

abstract type AbstractTreeNode{E,ID}

E is element type, ID is id type

Missing docstring.

Missing docstring for get_parent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_children. Check Documenter's build log for details.

Missing docstring.

Missing docstring for rem_parent!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_parent!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for has_child. Check Documenter's build log for details.

Missing docstring.

Missing docstring for has_parent. Check Documenter's build log for details.

GraphUtils.validate_tree — Function

validate_tree(n::AbstractTreeNode)

Ensure that the transform tree is in fact a tree–no cycles, and no duplicate ids

GraphUtils.validate_embedded_tree — Function

validate_embedded_tree(graph,f=v->get_node(graph,v))

Verify that all graph edges are mirrored by the parent-child structure stored in the nodes.

Factory World

GraphUtils.construct_vtx_map — Function

construct_vtx_map(vtxs,dims)

Returns a matrix M such that MV[i,j] == v, where v is the index of the vertex whose coordinates are (i,j)

Arguments:

vtxs : a list of integer coordinates

dims : the dimensions of the grid

GraphUtils.construct_edge_cache — Function

construct_edge_cache(vtxs,vtx_map)

Returns a cache such that cache[v] = {(0,1),(0,0),...}, the set of all valid directions in which a robot may move from vertex v.

Arguments:

vtxs : a list of integer coordinates

vtxmap : a matrix such that `vtxmap[i,j] = v, wherevtxs[i,j = v`

GraphUtils.construct_expanded_zones — Function

construct_expanded_zones(vtxs,vtx_map,pickup_zones,dropoff_zones;
    shapes=[(1,1),(1,2),(2,1),(2,2)])

A utility for constructing drop-off/pick-up zones where large-footprint objects need to be delivered. Returns a dictionary mapping vertices to a dict of shape=>vtxs for expanded size delivery zones. For each starting vertex, the expanded zone is selected as the appropriately sized region that overlaps with the original vertex, does not overlap with any obstacles, and has minimal overlap with all other vertices in the zone list.

Arguments:

vtxs : a list of integer coordinates

vtxmap : a matrix such that `vtxmap[i,j] == v, wherevtxs[v] == (i,j)`

zones : a list of integer vertex indices identifying the zones to be expanded

GraphUtils.validate_expanded_zones — Function

validate_expanded_zones(vtx_map,expanded_zones)

Verify that all vertices of each expanded zone do not overlap with obstacle regions, and that each expanded zone contains the original vertex.

GraphUtils.SparseDistanceMatrix — Type

SparseDistanceMatrix{G,M}

Stores a graph and sparse distance matrix. When the distance matrix is queried for the distance between start and goal, it first checks that this distance has been computed. If not, it will first compute the distances from all nodes to goal before returning the queried distance.

GraphUtils.remap_idx — Function

remap_idx(m::SparseDistanceMatrix,v,config)

Maps an index and configuration from the base grid to the transformed grid represented by m.graph. config represents the position of the query pt relative to the coordinates that correspond to v.

Missing docstring.

Missing docstring for get_val_and_status. Check Documenter's build log for details.

GraphUtils.recompute_cached_distances! — Function

recompute_cached_distances!

Recompute all cached distances (important if, e.g., the graph has been modified)

Missing docstring.

Missing docstring for remove_edges!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_edges!. Check Documenter's build log for details.

GraphUtils.config_index_to_tuple — Function

config_index_to_tuple(shape::Tuple{Int,Int}, idx::Int)

Convert a config index to a tuple coordinate representation. Assumes the following layout for config indices given shape (m,n): 1 2 ... n ____________________________ 1 | 1 2 ... n 2 | n+1 n+2 ... 2n ... | m | (m-1)n+1 (m-1)n+2 ... m*n

Missing docstring.

Missing docstring for config_tuple_to_index. Check Documenter's build log for details.

GraphUtils.RemappedDistanceMatrix — Type

RemappedDistanceMatrix

Stores a distance matrix that corresponds to a "shifted" grid graph with a new configuration space layered over a base graph. The shape m.s of m::RemappedDistanceMatrix defines the convolution kernel used to generate the modified.

Missing docstring.

Missing docstring for `GridFactoryEnvironment,

get_graph

getxdim`. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_y_dim. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_cell_width. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_transition_time. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_vtxs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_pickup_zones. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_dropoff_zones. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_obstacles. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_pickup_vtxs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_dropoff_vtxs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_obstacle_vtxs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_num_free_vtxs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_free_zones. Check Documenter's build log for details.

Sorting

module=[sorting]